backend/requirements.txt
asgiref==3.2.10
Django==3.1.14
django-colorfield==0.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
gunicorn==20.0.4
Pillow==9.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dotenv==0.21.0
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1